Add fixed Scilab-rl test environment
